explicit cast

+++ firmware/controllers/actuators/gppwm/gppwm_channel.cpp
@@ -45,13 +45,13 @@ expected<float> readGppwmChannel(gppwm_channel_e channel) {
 	case GPPWM_AuxLinear2:
 		return Sensor::get(SensorType::AuxLinear2);
 	case GPPWM_GppwmOutput1:
-		return engine->outputChannels.gppwmOutput[0];
+		return (float)engine->outputChannels.gppwmOutput[0];
 	case GPPWM_GppwmOutput2:
-		return engine->outputChannels.gppwmOutput[1];
+		return (float)engine->outputChannels.gppwmOutput[1];
 	case GPPWM_GppwmOutput3:
-		return engine->outputChannels.gppwmOutput[2];
+		return (float)engine->outputChannels.gppwmOutput[2];
 	case GPPWM_GppwmOutput4:
-		return engine->outputChannels.gppwmOutput[3];
+		return (float)engine->outputChannels.gppwmOutput[3];
 	}
 
 	return unexpected;
